+++ packages/widgetbook/lib/src/core/args/int_arg.dart
@@ -1,14 +1,37 @@
 import '../fields/fields.dart';
 import '../framework/arg.dart';
 
-class IntArg extends Arg<int> with SingleFieldOnly {
+class IntArg extends StyledArg<int, IntArgStyle> with SingleFieldOnly {
   IntArg(
     super.value, {
     super.name,
+    super.style = const InputIntArgStyle(),
   });
 
   @override
-  Field<int> get field {
+  Field<int> get field => style.toField(value);
+}
+
+class NullableIntArg extends StyledArg<int?, IntArgStyle> with SingleFieldOnly {
+  NullableIntArg(
+    super.value, {
+    super.name,
+    super.style = const InputIntArgStyle(),
+  });
+
+  @override
+  Field<int> get field => style.toField(value ?? 0);
+}
+
+sealed class IntArgStyle extends ArgStyle<int> {
+  const IntArgStyle();
+}
+
+final class InputIntArgStyle extends IntArgStyle {
+  const InputIntArgStyle();
+
+  @override
+  Field<int> toField(int value) {
     return IntInputField(
       name: 'value',
       initialValue: value,
@@ -16,17 +39,25 @@ class IntArg extends Arg<int> with SingleFieldOnly {
   }
 }
 
-class NullableIntArg extends Arg<int?> with SingleFieldOnly {
-  NullableIntArg(
-    super.value, {
-    super.name,
+final class SliderIntArgStyle extends IntArgStyle {
+  const SliderIntArgStyle({
+    required this.min,
+    required this.max,
+    required this.divisions,
   });
 
+  final int min;
+  final int max;
+  final int divisions;
+
   @override
-  Field<int> get field {
-    return IntInputField(
+  Field<int> toField(int value) {
+    return IntSliderField(
       name: 'value',
-      initialValue: value ?? 0,
+      initialValue: value,
+      min: min,
+      max: max,
+      divisions: divisions,
     );
   }
 }