+++ packages/widgetbook/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **FEAT**: Add `style` to `IntArg` and `DoubleArg` to support different input styles (e.g. sliders). ([#1792](https://github.com/widgetbook/widgetbook/pull/1792))
+
 ## 4.0.0-alpha.5
 
 - **BREAKING**: Merge `widgetbook_generator` and `widgetbook_test` packages into `widgetbook` package. ([#1772](https://github.com/widgetbook/widgetbook/pull/1772))